hw4/4-1/pyproject.toml
[tool]
[tool.pdm]

[project]
name = ""
version = ""
description = ""
authors = [
    {name = "alan910127", email = "alan469357@gmail.com"},
]
dependencies = [
    "fastapi>=0.88.0",
    "uvicorn>=0.20.0",
    "python-multipart>=0.0.5",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
]
requires-python = ">=3.10"
license = {text = "MIT"}

[tool.pdm.scripts]
dev = "uvicorn src.main:app --host '0.0.0.0' --port 9000 --reload"
